Seperating factory classes
